:sparkles: New Features: Output handler
Congrats! I have got the chance to use this project during my job Interview, and it works for two algorithms problems. However, when I was trying to deal with list-nodes, it caused a memory leak problem.

Added: when type of input == 'listnode', return an Array with all nodes' values;

Added: when type of input == 'treenode', return a Binary tree's pre-order values